Fix compile errors, fix some style and remove dons NDEF

+++ PN532/snep.cpp
@@ -80,7 +80,10 @@ int16_t SNEP::read(uint8_t *buf, uint8_t len, uint16_t timeout)
 	}
 
 	// check message's length
-	uint32_t length = (buf[2] << 24) + (buf[3] << 16) + (buf[4] << 8) + buf[5];
+	// TODO: fix this, surely 8 bits, shifted 24 times is undefined?
+	uint32_t length = (buf[2] << (uint8_t)24) + 
+					  (buf[3] << (uint8_t)16) + 
+					  (buf[4] << (uint8_t)8) + buf[5];
 	// length should not be more than 244 (header + body < 255, header = 6 + 3 + 2)
 	if (length > (status - 6)) {
 		DMSG("The SNEP message is too large: ");
